Fixtures preparation
* use small shops fixture
* extract users loading code for better reuse
