+++ features/management/commands/setup_e2e_data.py
@@ -1,30 +1,10 @@
 from django.conf import settings
 from django.core.management import call_command
 from django.core.management.base import BaseCommand
-from features.utils.fixtures import json
-from features.utils.auth.account_handling import create_database_user
+from features.utils.fixtures import json, database_loader
 from saleor.account.models import User
 
 import os
-
-
-def load_users_in_database(variant):
-    users = []
-    for persona in 'Consommateurs', 'Producteurs', 'Responsables', 'Rex', 'Softozor':
-        user_data = json.load(os.path.join(
-            settings.FIXTURE_DIRS[0], variant, 'Users', persona + '.json'))
-        if isinstance(user_data, list):
-            for user in user_data:
-                stored_user = create_database_user(user)
-                assert stored_user.pk == user['id'], '[%s] - %d should equal %d' % (
-                    persona, stored_user.pk, user['id'])
-                users.append(user)
-        else:
-            stored_user = create_database_user(user_data)
-            assert stored_user.pk == user_data['id'], '[%s] - %d should equal %d' % (
-                persona, stored_user.pk, user_data['id'])
-            users.append(user_data)
-    return users
 
 
 class Command(BaseCommand):
@@ -40,7 +20,7 @@ class Command(BaseCommand):
         output_folder = options['user_output_folder']
         variant = options['fixture_variant']
 
-        users = load_users_in_database(variant)
+        users = database_loader.load_users_in_database(variant)
         json.dump(users, os.path.join(output_folder, variant, 'Users.json'))
 
         call_command('loaddata', os.path.join(
