+++ features/steps/consumerVisitsShop.py
@@ -40,9 +40,7 @@ def step_impl(context):
     context.test.assertEqual(context.expected_shop_list, context.response)
 
 
-# TODO: complete data_factory with more data like latitude / longitude, name, description when they are ready
 # TODO: instead of trying to get the data from the database, get it from the graphql response fixture!
-# TODO: maybe generate the response here based on the fixture database Shops.json
 @then(u'il obtient la liste de tous les Produits qui y sont publiés')
 def step_impl(context):
     current_shop = context.current_shop
